cli: Treat an empty syncfolder as not configured
My `syncfolder` is set to an empty string.
I believe I had sync configured once and removed the folder since,
leaving me with an empty string.

This PR ignores empty strings and treats them as unconfigured.
Without that PR, the `workspaces` folder would be resolved
relative to the current dir, leading to file-not-found errors down the line.