cli: Support workspaces with `--package` flag
Previously `powerpack` only worked with the root package of a workspace.
You can now have multiple packages in a workspace and use the
`--package` flag to specify which package to build, link or package. The
`workflow/` directory containing the package information must be in the
same directory as the manifest file for the particular package.